Break PacmanSupporter to 3 separate files
